NEW Allow install in non-dev modes
While this is a dev tool, it's also a good way to allow developers to play with a GraphQL API in a SilverStripe demo environment.
This environment should not run in dev mode, but rather enforce specific access control for this controller.

I've changed the default installation instructions to "require --dev", which should provide a first line of defence.
Developers installing this tool need to opt-in to making it available outside of dev contexts (assuming they're using composer properly).

The next level of defence is permission checks, rather than hard wiring the tool to only work in dev mode.
These checks are enforced by the DevelopmentAdmin parent controller.

The side effect of this is that every GraphQL request in non-dev mode will carry a session, making it harder to simulate
unauthenticated GraphQL responses. I don't think that's a problem in the module as such, but rather a shortcoming in how we've set up GraphQL.
It should use OAuth, JWT or API keys, and allow setting those through GraphiQL.
Given that it's a non issue in dev mode, I'm fine with that change.

Aside: Nesting controllers and relying on access control in parent controllers is a bit dangerous.
You could hook this controller to a different route, and accidentally run it without these checks.
That's not a problem specific to this module though.